refactor main UI folder for streamlit
