+++ .vscode/settings.json
@@ -11,5 +11,8 @@
       "tests"
     ],
     "python.testing.unittestEnabled": false,
-    "python.testing.pytestEnabled": true
+    "python.testing.pytestEnabled": true,
+    "python.analysis.extraPaths": [
+      "./gpt_nexus/streamlit_"
+    ]
   }
