+++ gpt_nexus/streamlit_ui/actions.py
@@ -1,6 +1,6 @@
 import streamlit as st
 
-from gpt_nexus.ui.cache import get_nexus
+from gpt_nexus.streamlit_ui.cache import get_nexus
 
 
 def actions_page(username, win_height):
