+++ gpt_nexus/streamlit_ui/agent.py
@@ -1,6 +1,6 @@
 import streamlit as st
 
-from gpt_nexus.ui.cache import get_nexus
+from gpt_nexus.streamlit_ui.cache import get_nexus
 
 
 def save_new_agent(name, agent_type):
