+++ gpt_nexus/streamlit_ui/assistants_panel.py
