+++ gpt_nexus/streamlit_ui/memory.py
@@ -1,8 +1,8 @@
 import streamlit as st
 
 from gpt_nexus.nexus_base.nexus_models import MemoryType
-from gpt_nexus.ui.cache import get_nexus
-from gpt_nexus.ui.embeddings import get_agent, view_embeddings
+from gpt_nexus.streamlit_ui.cache import get_nexus
+from gpt_nexus.streamlit_ui.embeddings import get_agent, view_embeddings
 
 
 def add_memory_to_store(chat, memory_store):
